refactor(website): extract summer update components
