+++ apps/networkcanvas.com/components/summer-update/summerUpdateMotion.ts
@@ -1,0 +1,12 @@
+export const summerUpdateEasing = [0.22, 1, 0.36, 1] as [
+  number,
+  number,
+  number,
+  number,
+];
+
+export const summerUpdateRevealMotion = {
+  distance: 32,
+  duration: 0.9,
+  easing: summerUpdateEasing,
+} as const;
